+++ src/component/feature/ProductContainer/ProductContainer.tsx
@@ -1,25 +1,48 @@
+import { postCartItem } from "../../../api/cartItem";
+import { useCartContext } from "../../../hook/CartContext";
 import { CartItemType } from "../../../types/cartItem";
 import { ProductType } from "../../../types/product";
+import Button from "../../unit/Button/Button";
+import { QuantitySelector } from "../../unit/QuantitySelector/QuantitySelector";
 import Product from "../Product/Product";
 import { ProductContainerLayout } from "./ProductContainer.style";
 
 interface ProductContainerProps {
   products: ProductType[];
-  cartItemList: CartItemType[];
-  onChange: () => void;
 }
 
-export default function ProductContainer({
-  products,
-  cartItemList,
-  onChange,
-}: ProductContainerProps) {
+export default function ProductContainer({ products }: ProductContainerProps) {
+  const { cartItemList, dispatch } = useCartContext();
+
+  const onChange = () => {
+    dispatch({ type: "update" });
+  };
+
+  const AddToCartButton = (id: number) => {
+    const handleClick = async () => {
+      await postCartItem({ productId: id, quantity: 1 });
+      onChange();
+    };
+
+    return (
+      <Button onClick={handleClick}>
+        <img src="./add-shopping-cart.svg" />
+        <p>담기</p>
+      </Button>
+    );
+  };
+
   return (
     <div css={ProductContainerLayout}>
       {products.map((product) => {
         const selectedCardItems = cartItemList.filter(
           (cartItem: CartItemType) => Number(product.id) === cartItem.product.id
         );
+        const isSelected = selectedCardItems.length !== 0;
+        const cartProduct = cartItemList.filter(
+          (cartItem) => cartItem.product.id === product.id
+        );
+
         return (
           <Product
             key={product.id}
@@ -27,9 +50,17 @@ export default function ProductContainer({
             imageUrl={product.imageUrl}
             name={product.name}
             price={product.price}
-            selectedCartItems={selectedCardItems}
-            onChange={onChange}
-          />
+          >
+            {isSelected ? (
+              <QuantitySelector
+                quantity={cartProduct[0].quantity}
+                cartId={cartProduct[0].id}
+                onChange={onChange}
+              />
+            ) : (
+              AddToCartButton(product.id)
+            )}
+          </Product>
         );
       })}
     </div>
